messages/notify/pyproject.toml
[tool.poetry]
name = "notify"
version = "0.1.0"
description = ""
authors = ["Marcin Tyszkiewicz <56219944+mtyszkiewicz@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
paho-mqtt = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
